requirements.txt
astroid==2.0.4
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
elasticsearch==6.3.1
falcon==1.4.1
google-api-core==1.5.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-datastore==1.7.1
googleapis-common-protos==1.5.3
grpcio==1.16.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
protobuf==3.6.1
pyasn1-modules==0.2.2
pyasn1==0.4.4
pylint==2.1.1
python-mimeparse==1.6.0
pytz==2018.7
requests==2.20.0
rsa==4.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.24
wrapt==1.10.11
